move jck expr to the right location, move private classes to separate files
