+++ jaxb-tck/src/tools/jttck/jck.lib/src/main/java/com/sun/jck/lib/LessEqualExpr.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2026 Contributors to the Eclipse Foundation.
+ * Copyright (c) 2002, 2018 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * https://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+package com.sun.jck.lib;
+
+import com.sun.javatest.TestEnvironment;
+
+class LessEqualExpr extends BinaryExpr {
+    LessEqualExpr(Expr left, Expr right) {
+        super(left, right);
+    }
+
+    public String eval(TestEnvironment e) throws Fault {
+        return String.valueOf(left.evalInt(e) <= right.evalInt(e));
+    }
+
+    int precedence() {
+        return PREC_LE;
+    }
+
+    public String toString() {
+        return "`" + left + "<=" + right + "'";
+    }
+}
