+++ jaxb-tck/src/tools/jttck/jck.lib/src/main/java/com/sun/jck/lib/NameExpr.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2026 Contributors to the Eclipse Foundation.
+ * Copyright (c) 2002, 2018 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0, which is available at
+ * https://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the
+ * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+ * version 2 with the GNU Classpath Exception, which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ */
+package com.sun.jck.lib;
+
+import com.sun.javatest.TestEnvironment;
+import com.sun.javatest.util.StringArray;
+
+class NameExpr extends Expr {
+    NameExpr(String key) {
+        this.key = key;
+    }
+
+    public String eval(TestEnvironment e) throws Fault {
+        try {
+            String[] v = (e == null ? null : e.lookup(key));
+            if (v == null) {
+                throw new Fault("name not defined: " + key);
+            }
+            return StringArray.join(v);
+        } catch (TestEnvironment.Fault ex) {
+            throw new Fault("error looking up: " + key);
+        }
+    }
+
+    int precedence() {
+        return PREC_TRM;
+    }
+
+    public String toString() {
+        return key;
+    }
+
+    private final String key;
+}
